Move `StoppingPower` definition to the atomic module
